#1546: split filter engine into responsibility-scoped submodules
Closes #1546.

Split userspace-dp/src/filter/engine.rs (1247 LOC, 51 fns) into five
responsibility-scoped submodules under userspace-dp/src/filter/engine/:

  matching.rs       — term_matches family (V4/V6, allocation-free,
                      #[inline(always)] preserved)
  eval.rs           — input/output filter evaluation (non-tx) plus
                      the routing-instance evaluator and its precheck
                      `interface_filter_affects_route_lookup`
  tx_selection.rs   — TX-selection runtime lookup (per-packet hot
                      path with policer application)
  cache_sensitive.rs — cached TX-selection reconstruction +
                      DSCP/semantics comparison helpers
  policer.rs        — three-color policer application
                      (apply_term_three_color_policer + cached form)

Pure code motion. Function bodies are byte-identical with the
pre-split engine.rs; #[inline] / #[inline(always)] attributes are
preserved verbatim. Visibility tightened from engine-wide pub(crate)
to pub(super) for engine-local helpers, with pub(crate) re-exports at
engine/mod.rs maintaining the existing call surface in the rest of
the crate. filter/policer.rs (the standalone three-color runtime,
distinct from the new engine/policer.rs) is untouched.

Plan history:
  v1 (3c365bcb)  drafted
  v2 (a5e8acb6)  addressed Codex r1 PLAN-NEEDS-MAJOR + AGY r1
                 PLAN-NEEDS-MINOR — rustc/inlining wording corrected
                 to cite Rust Reference + rustc dev guide + ThinLTO;
                 policer concurrency invariant rewritten (Mutex, not
                 atomic CAS); interface_filter_affects_route_lookup
                 moved cache_sensitive -> eval; tests use sibling
                 *_tests.rs pattern per modularity discipline.
  v2 reviews: AGY r2 PLAN-READY; Codex r2 infra-blocked (CreateProcess
              failures across 2 retries) — retried; AGY r2 endorsed
              the implementation directly via worktree inspection.

#1546 AC2 ("Cache-sensitive comparison helpers live in one module
with tests covering add, remove, same-ifindex content change,
positional-ID stability, and policer runtime-shape changes") satisfied:

  - same-ifindex content change: filter/tests.rs:1733 (pre-existing)
  - positional-ID stability:     filter/tests.rs:1806 (pre-existing)
  - policer runtime-shape:       filter/tests.rs:1851 (pre-existing)
  - filter added (NEW):          filter/tests.rs:1921
  - filter removed (NEW):        filter/tests.rs:1958

Validation:
  - cargo build clean
  - cargo test --release --bin xpf-userspace-dp: 1417/1417 pass
  - 5x flake check on filter:: scope: 38/38 pass each iteration
  - go test ./...: clean
  - Pre-existing master test failure (snat_contract_doc_guard)
    confirmed unrelated to this change.
