+++ docs/pr/1546-filter-engine-split/reviewer-ids.md
@@ -4,8 +4,13 @@ Recorded so long-running agents can fetch verdicts by ID even if
 Codex/AGY session state evicts.
 
 ## Plan review round 1
-- Codex: <pending>
-- AGY:   <pending>
+- Codex: task-mpmyrfos-68aj5c — PLAN-NEEDS-MAJOR (4 findings)
+- AGY:   adversarial-review-mpmyryoh-ekjl7p — PLAN-NEEDS-MINOR (test placement)
+
+## Plan review round 2
+- Codex: task-mpmz9amw-ib4n8w — infra failure (CreateProcess errors)
+- Codex retry: task-mpmzi3li-e6irdd
+- AGY:   adversarial-review-mpmz9ii5-50whwe — PLAN-READY
 
 ## Code review round 1
 - Codex: <pending>
